Resolute AI Automation
Hi,
Only Login functionality can be automated as elements are present in different frames in the web app. But these frames is not recognised as iframe in xpath. Therefore element is always hidden or masked. To prove this i found xpath of the main frame "Dashboard" present in dashboard/Home.